Fix MLAWriteCallback prototype for partial writes
The new prototype allows returning an error code, and also to
return a "ok but write completed only partially", which the
Rust runtime will handle for you.

It also handles write requests larger than 4GB, which are not
handled by some OSes APIs, and will be chunked for us by the
Rust runtime (the safest option for us to avoid arithmetic
mistakes).