Fix #226
CompressionFailSafeReader was returning Ok(0) too early when BrotliDecompressStream was returning a NeedsMoreInput with an output_offset of 0.

We fix this by checking this case, getting more input if needed and retrying decompression.

All sorts of other corner cases have been handled in this new implementation, including ones potentially not handled by previous implementation with BrotliDecompressor.

e.g. it did not handle the case where when BrotliDecompressor calls BrotliDecompressStream and it returns NeedsMoreInput twice, it could return an error if no input was consumed.

We had to return to a potentially slower implementation where we generally call BrotliDecompressStream by feeding it byte by byte. Otherwise it may not return all its output on failure.

This commits also removes unneeded enum states in CompressionFailSafeReader and moves handling of BrotliDecompressStream in a dedicated brotli_decompress_stream function.